backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
crewai
langchain-community
duckduckgo-search
langchain-openai
python-dotenv